feat: 주문 정보 담는 값 객체
